refactor(chat): share anonymization pipeline core
