+++ apps/web/package.json
@@ -29,6 +29,7 @@
     "@libpdf/core": "catalog:",
     "@posthog/react": "^1.9.0",
     "@pydantic/genai-prices": "^0.0.56",
+    "@stll/anonymize-chat": "workspace:*",
     "@stll/anonymize-data": "catalog:",
     "@stll/anonymize-wasm": "catalog:",
     "@stll/api": "workspace:*",
